fix: fix missing copy button
