+++ packages/kit/src/components/TabPageHeader/HeaderTitle.tsx
@@ -1,45 +1,25 @@
 import { useMemo } from 'react';
 
-import { XStack, useIsHorizontalLayout } from '@onekeyhq/components';
+import platformEnv from '@onekeyhq/shared/src/platformEnv';
 import { EAccountSelectorSceneName } from '@onekeyhq/shared/types';
 
-import { UrlAccountNavHeader } from '../../views/Home/pages/urlAccount/UrlAccountNavHeader';
-import {
-  AccountSelectorActiveAccountHome,
-  AccountSelectorProviderMirror,
-  NetworkSelectorTriggerHome,
-} from '../AccountSelector';
+import { AccountSelectorProviderMirror } from '../AccountSelector';
+
+import { UrlAccountPageHeader } from './urlAccountPageHeader';
 
 export function HeaderTitle({
   sceneName,
 }: {
   sceneName: EAccountSelectorSceneName;
 }) {
-  const isHorizontal = useIsHorizontalLayout();
-
   const item = useMemo(() => {
-    if (sceneName === EAccountSelectorSceneName.homeUrlAccount) {
-      return (
-        <XStack gap="$2.5" ai="center">
-          <UrlAccountNavHeader.Address key="urlAccountNavHeaderAddress" />
-          {isHorizontal ? (
-            <NetworkSelectorTriggerHome
-              num={0}
-              recordNetworkHistoryEnabled
-              hideOnNoAccount
-            />
-          ) : null}
-          <AccountSelectorActiveAccountHome
-            num={0}
-            showAccountAddress={false}
-            showCopyButton
-            showCreateAddressButton={false}
-            showNoAddressTip={false}
-          />
-        </XStack>
-      );
+    if (
+      !platformEnv.isNativeIOS &&
+      sceneName === EAccountSelectorSceneName.homeUrlAccount
+    ) {
+      return <UrlAccountPageHeader />;
     }
-  }, [isHorizontal, sceneName]);
+  }, [sceneName]);
   return (
     <AccountSelectorProviderMirror
       enabledNum={[0]}
